+++ nata-monkey/src/main/java/com/nata/state/Bag.java
@@ -1,0 +1,51 @@
+package com.nata.state;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-13 16:53
+ */
+
+import java.util.Iterator;
+public class Bag<Item> implements Iterable<Item> {
+
+    private Node first;
+
+    private class Node{
+        Item item;
+        Node next;
+    }
+
+    public void add(Item item) {
+        Node oldFirst = first;
+        first = new Node();
+        first.item =item;
+        first.next = oldFirst;
+    }
+
+    public boolean isEmpty() {
+        return first == null;
+    }
+
+    @Override
+    public Iterator<Item> iterator() {
+        return new ListIterator();
+    }
+
+    private class ListIterator implements Iterator<Item>{
+        private  Node current = first;
+        @Override
+        public boolean hasNext() {
+            return  current != null;
+        }
+
+        @Override
+        public Item next() {
+            Item item = current.item;
+            current = current.next;
+            return item;
+        }
+    }
+
+}
+