+++ nata-monkey/src/main/java/com/nata/state/Graph.java
@@ -1,0 +1,36 @@
+package com.nata.state;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-13 17:17
+ */
+public class Graph {
+    private final int V;
+    private int E;
+    private Bag<Integer>[] adj;
+
+
+    public Graph(int V){
+        this.V = V;
+        this.E = 0;
+        adj = (Bag<Integer>[]) new Bag[V];
+        for(int v = 0; v < V; v++){
+            adj[v] = new Bag<>();
+        }
+    }
+
+    public int V(){
+        return V;
+    }
+
+    public int E(){
+        return E;
+    }
+
+    public void addEdge(int v,int w){
+        adj[v].add(w);
+        adj[w].add(v);
+        E++;
+    }
+}