+++ CHANGELOG.md
@@ -92,3 +92,8 @@
 
 ### (2016/4/11)
 - Rewrite rules structure
+
+### (2016/4/13)
+- Add DFS Monkey
+- Add StateGraph to manage the states and action relationships
+- Add ActionEdge to link two states